レビュー指摘対応
- `LocalizedStringDefinition` のCloneableを削除
- プラガブル化をControllerを使用する方式に変更
- service-configの設定を整理
- 新規ファイルにcopyrightを表記
- 変数名調整
