+++ iplass-admin/src/main/java/org/iplass/adminconsole/client/metadata/ui/template/TemplateEditPaneControllerImpl.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2024 DENTSU SOKEN INC. All Rights Reserved.
+ * 
+ * Unless you have purchased a commercial license,
+ * the following license terms apply:
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.iplass.adminconsole.client.metadata.ui.template;
+
+import org.iplass.mtp.web.template.definition.TemplateDefinition;
+
+public class TemplateEditPaneControllerImpl implements TemplateEditPaneController {
+
+	@Override
+	public void beforeUpdate(boolean updateRpc, TemplateDefinition definition) {
+		// 何もしない
+	}
+
+}
