+++ iplass-core/src/main/java/org/iplass/mtp/definition/LocalizedStringDefinition.java
@@ -48,13 +48,4 @@ public class LocalizedStringDefinition implements Serializable, Cloneable {
 		this.stringValue = stringValue;
 	}
 
-	public Object clone() {
-		// Object#clone()はGWTでサポートされていないため、自前でcloneする
-		LocalizedStringDefinition clone = new LocalizedStringDefinition();
-
-		clone.setLocaleName(localeName);
-		clone.setStringValue(stringValue);
-
-		return clone;
-	}
 }
